+++ src/galeforce/actions/action.ts
@@ -63,85 +63,99 @@ export default class Action<TResult> {
 
         actionDebug(`${chalk.bold.magenta(this.payload._id)} | ${chalk.bold.yellow('execute')} \u00AB %O`, this.payload);
         try {
+            // Ensure required payload properties exist
             if (typeof this.payload.endpoint === 'undefined') {
                 throw new Error('[galeforce]: Action endpoint is required but undefined.');
             }
             if (typeof this.payload.method === 'undefined') {
                 throw new Error('[galeforce]: Action method is required but undefined.');
             }
-
-            // Execute-time property checks
-
-            if (this.payload.gameName && !this.payload.tagLine) {
-                throw new Error('[galeforce]: .gameName() must be chained with .tagLine().');
-            }
-
-            // Increment rate limit for non-Data Dragon requests.
             if (this.payload.type && ['lol', 'val', 'riot'].includes(this.payload.type) && typeof this.payload.region === 'undefined') {
                 throw new Error('[galeforce]: Action payload region is required but undefined.');
             }
 
+            // Runtime property checks to ensure the Action payload is properly formed
+            if (this.payload.gameName && !this.payload.tagLine) {
+                throw new Error('[galeforce]: .gameName() must be chained with .tagLine().');
+            }
+
+            // Set request depending on type
             let request: Request;
-            if (this.payload.type === 'gc') {
+            switch (this.payload.type) {
+            case 'gc': // Game Client requests (requiring an SSL certificate)
                 request = this.submodules.RiotAPI.gcRequest(
                     this.payload.endpoint,
                     this.payload,
                     this.payload.query,
                     this.payload.body,
                 );
-            } else if (this.payload.type === 'ddragon-buffer') {
+                break;
+            case 'ddragon-buffer': // Data Dragon images
                 request = this.submodules.RiotAPI.bufferRequest(
                     this.payload.endpoint,
                     this.payload,
                     this.payload.query,
                     this.payload.body,
                 );
-            } else {
+                break;
+            default: // All other request types
                 request = this.submodules.RiotAPI.request(
                     this.payload.endpoint,
                     this.payload,
                     this.payload.query,
                     this.payload.body,
                 );
+                break;
             }
 
+            // Schedule request depending on HTTP method
             let response: TResult;
-
-            if (this.payload.method === 'GET') {
+            switch (this.payload.method) {
+            case 'GET':
+                // Schedule a GET request using the selected rate limiter
                 response = (await this.submodules.RateLimiter.schedule(() => request.get(), this.payload.region)).data as TResult;
-            } else if (this.payload.method === 'POST') {
+                break;
+            case 'POST':
+                // Ensure that the POST request has a defined body
                 if (typeof this.payload.body === 'undefined') {
                     throw new Error('[galeforce]: Action payload body is required but undefined.');
                 }
+                // Schedule a POST request using the selected rate limiter
                 response = (await this.submodules.RateLimiter.schedule(() => request.post(), this.payload.region)).data as TResult;
-            } else if (this.payload.method === 'PUT') {
+                break;
+            case 'PUT':
+                // Ensure that the PUT request has a defined body
                 if (typeof this.payload.body === 'undefined') {
                     throw new Error('[galeforce]: Action payload body is required but undefined.');
                 }
+                // Schedule a PUT request using the selected rate limiter
                 response = (await this.submodules.RateLimiter.schedule(() => request.put(), this.payload.region)).data as TResult;
-            } else {
+                break;
+            default: // Should never occur, but throw an error if reached
                 throw new Error('[galeforce]: Invalid action method provided.');
             }
 
             actionDebug(`${chalk.bold.magenta(this.payload._id)} | ${chalk.bold.yellow('return')} \u00AB ${chalk.bold.green(200)}`);
             return response;
         } catch (e) {
-            if (e.response?.status) {
-                actionDebug(`${chalk.bold.magenta(this.payload._id)} | ${chalk.bold.yellow('return')} \u00AB ${chalk.bold.red(e.response.status)}`);
-                if (e.response.status === 401) {
-                    throw new Error('[galeforce]: No Riot API key was provided. Please ensure that your key is present in your configuration file or object. (401 Unauthorized)');
-                } else if (e.response.status === 403) {
-                    throw new Error('[galeforce]: The provided Riot API key is invalid or has expired. Please verify its authenticity. (403 Forbidden)');
-                } else {
-                    throw new Error(`[galeforce]: Data fetch failed with status code ${e.response.status}`);
-                }
+            actionDebug(`${chalk.bold.magenta(this.payload._id)} | ${chalk.bold.yellow('return')} \u00AB ${chalk.bold.red(e.response?.status || 'error')}`);
+            switch (e.response?.status) {
+            case 401:
+                // Handle HTTP 401 Unauthorized errors, returned when no API key is used when requesting data.
+                throw new Error('[galeforce]: No Riot API key was provided. Please ensure that your key is present in your configuration file or object. (401 Unauthorized)');
+            case 403:
+                // Handle HTTP 403 Forbidden errors, caused by the use of invalid or expired API keys.
+                throw new Error('[galeforce]: The provided Riot API key is invalid or has expired. Please verify its authenticity. (403 Forbidden)');
+            case undefined:
+                throw e;
+            default:
+                // Generic HTTP error handling.
+                throw new Error(`[galeforce]: Data fetch failed with status code ${e.response.status}`);
             }
-
-            actionDebug(`${chalk.bold.magenta(this.payload._id)} | ${chalk.bold.yellow('return')} \u00AB ${chalk.bold.red('error')}`);
-            throw e;
         }
     }
 
+    /* eslint-disable class-methods-use-this */
     protected inferEndpoint(): void { /* Empty because this may be implemented by classes that inherit from Action */ }
 
     /**
@@ -154,6 +168,7 @@ export default class Action<TResult> {
     public URL(): string {
         this.inferEndpoint();
 
+        // Ensure required payload properties exist
         if (typeof this.payload.endpoint === 'undefined') {
             throw new Error('[galeforce]: Action endpoint is required but undefined.');
         }
@@ -165,6 +180,6 @@ export default class Action<TResult> {
             this.payload.body,
         );
 
-        return request.targetURL;
+        return request.targetURL; // Return only the target URL of the associated request
     }
 }