+++ src/galeforce/actions/data-dragon/scoreboard-icon-art.ts
@@ -5,12 +5,26 @@ import SubmoduleMap from '../../interfaces/submodule-map';
 export default class GetDataDragonScoreboardArt extends Action<Buffer> {
     constructor(submodules: SubmoduleMap, icon: 'champion' | 'items' | 'minion' | 'score' | 'spells') {
         super(submodules);
-        if (icon === 'champion') this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_CHAMPION_ART;
-        else if (icon === 'items') this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_ITEMS_ART;
-        else if (icon === 'minion') this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_MINION_ART;
-        else if (icon === 'score') this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_SCORE_ART;
-        else if (icon === 'spells') this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_SPELLS_ART;
-        else throw new Error('[galeforce]: Invalid constructor parameter provided.');
+
+        switch (icon) {
+        case 'champion':
+            this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_CHAMPION_ART;
+            break;
+        case 'items':
+            this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_ITEMS_ART;
+            break;
+        case 'minion':
+            this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_MINION_ART;
+            break;
+        case 'score':
+            this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_SCORE_ART;
+            break;
+        case 'spells':
+            this.payload.endpoint = ENDPOINTS.DATA_DRAGON.SCOREBOARD_ICONS_SPELLS_ART;
+            break;
+        default:
+            throw new Error('[galeforce]: Invalid constructor parameter provided.');
+        }
 
         this.payload.type = 'ddragon-buffer';
         this.payload.method = 'GET';