+++ src/riot-api/index.ts
@@ -13,6 +13,7 @@ import Tier from './enums/tiers';
 import Division from './enums/divisions';
 import Game from './enums/games';
 import Request from './requests';
+import { ConfigInterface } from '../galeforce/interfaces/config';
 
 const initDebug = debug('galeforce:init');
 
@@ -20,19 +21,18 @@ initDebug(`${chalk.bold('loading Game Client certificate chain')}`);
 const httpsAgent = new https.Agent({ ca: fs.readFileSync(path.join(__dirname, '..', '..', 'resource', 'riotgames.pem')) });
 
 export class RiotAPIModule {
-    private key: string;
+    private key?: string;
 
-    constructor(key: string) {
-        this.key = key;
+    constructor(options: ConfigInterface['riot-api']) {
+        this.key = options.key;
     }
 
     /**
      * @private
      *
-     * @param {String} template A list of string templates that are filled in with parameters.
-     * @param {Object} match Parameters to fill in variables for the string templates.
-     *
-     * @return {[String]} Substituted version of template with parameter values from match.
+     * @param template A list of string templates that are filled in with parameters.
+     * @param match Parameters to fill in variables for the string templates.
+     * @returns Substituted version of template with parameter values from match.
      */
     private static generateTemplateString(template: string, match: Record<string, unknown>): string {
         try {
@@ -48,13 +48,31 @@ export class RiotAPIModule {
         }
     }
 
+    /**
+     * Returns a general Riot API `Request` object, with the provided API key as a header.
+     *
+     * @param stringTemplate The URL string template to substitute `parameters` into.
+     * @param parameters The parameters to substitute into the URL `stringTemplate`.
+     * @param query The query to send along with the request.
+     * @param body The body of the request.
+     * @returns a `Request` object.
+     */
     public request(stringTemplate: string, parameters: Record<string, unknown>, query: object = {}, body: object = {}): Request {
         return new Request(RiotAPIModule.generateTemplateString(stringTemplate, parameters), body, {
             params: query,
-            headers: { 'X-Riot-Token': this.key },
+            headers: this.key ? { 'X-Riot-Token': this.key } : {},
         });
     }
 
+    /**
+     * Returns a `Request` object with a custom `https.Agent` containing the Riot Game Client SSL certificate chain.
+     *
+     * @param stringTemplate The URL string template to substitute `parameters` into.
+     * @param parameters The parameters to substitute into the URL `stringTemplate`.
+     * @param query The query to send along with the request.
+     * @param body The body of the request.
+     * @returns a `Request` object.
+     */
     public gcRequest(stringTemplate: string, parameters: Record<string, unknown>, query: object = {}, body: object = {}): Request {
         return new Request(RiotAPIModule.generateTemplateString(stringTemplate, parameters), body, {
             params: query,
@@ -62,6 +80,15 @@ export class RiotAPIModule {
         });
     }
 
+    /**
+     * Returns a `Request` object that provides the fetched data as an `ArrayBuffer`.
+     *
+     * @param stringTemplate The URL string template to substitute `parameters` into.
+     * @param parameters The parameters to substitute into the URL `stringTemplate`.
+     * @param query The query to send along with the request.
+     * @param body The body of the request.
+     * @returns a `Request` object.
+     */
     public bufferRequest(stringTemplate: string, parameters: Record<string, unknown>, query: object = {}, body: object = {}): Request {
         return new Request(RiotAPIModule.generateTemplateString(stringTemplate, parameters), body, {
             params: query,