+++ CHANGELOG.md
@@ -31,7 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
     ```javascript
     {
       'riot-api': {
-          key: '',
+          key: undefined,
       },
       'rate-limit': {
           type: 'bottleneck',
@@ -50,7 +50,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
       debug: [],
     }
     ```
-- Updated *galeforce:rate-limit* debugging output
+- Updated `galeforce:rate-limit` debugging output
 
 ### [0.3.0] (2021-05-03)